Next bit of NSInteger cleanup for -gui.
Fix all of the current warnings in Gorm so that it can be used to test
NSInteger things.



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gorm/trunk@36039 72102866-910b-0410-8b05-ffd578937521

+++ GormCore/GormClassEditor.m
@@ -1334,7 +1334,7 @@ shouldEditTableColumn: (NSTableColumn *)tableColumn
 
 @implementation GormClassEditor (NSBrowserDelegate)
 
-- (void)browser:(NSBrowser *)sender createRowsForColumn:(int)column inMatrix:(NSMatrix *)matrix
+- (void)browser:(NSBrowser *)sender createRowsForColumn:(NSInteger)column inMatrix:(NSMatrix *)matrix
 {
   NSArray      *classes = nil;
   NSEnumerator *en = nil;
