+++ GormPrefs/GormHeadersPref.m
@@ -14,7 +14,7 @@
 @end
 
 @implementation HeaderDataSource
-- (int) numberOfRowsInTableView: (NSTableView *)tv
+- (NSInteger) numberOfRowsInTableView: (NSTableView *)tv
 {
   NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
   NSArray *list = [defaults objectForKey: @"HeaderList"];
@@ -23,7 +23,7 @@
 
 - (id)          tableView: (NSTableView *)tv
 objectValueForTableColumn: (NSTableColumn *)tc
-	              row: (int)rowIndex
+	              row: (NSInteger)rowIndex
 {
   NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
   NSArray *list = [defaults objectForKey: @"HeaderList"];
@@ -123,14 +123,14 @@ objectValueForTableColumn: (NSTableColumn *)tc
 
 - (BOOL)    tableView: (NSTableView *)tableView
 shouldEditTableColumn: (NSTableColumn *)aTableColumn
-		  row: (int)rowIndex
+		  row: (NSInteger)rowIndex
 {
   BOOL result = NO;
   return result;
 }
 
 - (BOOL) tableView: (NSTableView *)tv
-   shouldSelectRow: (int)rowIndex
+   shouldSelectRow: (NSInteger)rowIndex
 {
   BOOL result = YES;
   return result;
