+++ GormPrefs/GormPalettesPref.m
@@ -36,7 +36,7 @@
 @end
 
 @implementation PaletteDataSource
-- (int) numberOfRowsInTableView: (NSTableView *)tv
+- (NSInteger) numberOfRowsInTableView: (NSTableView *)tv
 {
   NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
   NSArray *list = [defaults objectForKey: @"UserPalettes"];
@@ -45,7 +45,7 @@
 
 - (id)          tableView: (NSTableView *)tv
 objectValueForTableColumn: (NSTableColumn *)tc
-	              row: (int)rowIndex
+	              row: (NSInteger)rowIndex
 {
   NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
   NSArray *list = [defaults objectForKey: @"UserPalettes"];
@@ -116,14 +116,14 @@ objectValueForTableColumn: (NSTableColumn *)tc
 
 - (BOOL)    tableView: (NSTableView *)tableView
 shouldEditTableColumn: (NSTableColumn *)aTableColumn
-		  row: (int)rowIndex
+		  row: (NSInteger)rowIndex
 {
   BOOL result = NO;
   return result;
 }
 
 - (BOOL) tableView: (NSTableView *)tv
-   shouldSelectRow: (int)rowIndex
+   shouldSelectRow: (NSInteger)rowIndex
 {
   BOOL result = YES;
   return result;
